Factor wasm validator into a cpp file
Also small cleanup to CMake libraries
